Created Login Test Suite, moved Login tests  and Added Mochawesome, dot, allure dependencies
