+++ README.md
@@ -3,3 +3,6 @@ Zero Bank Testing Practice
 Added Babel file for Javascript formatisation
 Used Javascript Mocha as the framework
 Jenkins for CICI from local.
+
+Adding allure reporting:
+npm install @wdio/allure-reporter --save-dev
