+++ mochareports/wdio-0-0-mochawesome-reporter.log
@@ -1,0 +1,1 @@
+{"stats":{"suites":1,"tests":2,"passes":2,"pending":0,"failures":0,"start":"2020-06-22T22:33:00.862Z","end":"2020-06-22T22:33:10.121Z","duration":9259,"testsRegistered":2,"passPercent":100,"pendingPercent":0,"other":0,"hasOther":false,"skipped":0,"hasSkipped":false,"passPercentClass":"success","pendingPercentClass":"danger"},"suites":{"title":"","suites":[{"title":"E2E Tests - Help (chrome.83_0_4103_106.windows)","suites":[],"tests":[{"title":"Should Login to Application","fullTitle":"Should Login to Application","timedOut":false,"duration":5736,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"6361d9d7-e703-4618-a4e6-e47960660563","parentUUID":"b80f947b-f286-405d-ab03-007f02be9eab","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"86c11b1fa2b3ba1b08c75dad208d9f9c\"}]","state":"passed","err":{}},{"title":"should load help page","fullTitle":"should load help page","timedOut":false,"duration":2116,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"4f54a237-9179-4505-9ade-3bc82a48306b","parentUUID":"b80f947b-f286-405d-ab03-007f02be9eab","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"86c11b1fa2b3ba1b08c75dad208d9f9c\"}]","state":"passed","err":{}}],"pending":[],"root":false,"fullFile":"","file":"","passes":["6361d9d7-e703-4618-a4e6-e47960660563","4f54a237-9179-4505-9ade-3bc82a48306b"],"failures":[],"skipped":[],"duration":7856,"_timeout":0,"uuid":"b80f947b-f286-405d-ab03-007f02be9eab","beforeHooks":[],"afterHooks":[]}],"tests":[],"pending":[],"root":true,"fullFile":"","file":"","passes":[],"failures":[],"skipped":[],"duration":0,"_timeout":0,"uuid":"73197b6d-7f0b-463e-b207-9d45115eb56b","beforeHooks":[],"afterHooks":[]},"copyrightYear":2020}
