+++ mochareports/wdio-0-1-mochawesome-reporter.log
@@ -1,0 +1,1 @@
+{"stats":{"suites":1,"tests":3,"passes":3,"pending":0,"failures":0,"start":"2020-06-22T22:33:01.034Z","end":"2020-06-22T22:33:13.097Z","duration":12063,"testsRegistered":3,"passPercent":100,"pendingPercent":0,"other":0,"hasOther":false,"skipped":0,"hasSkipped":false,"passPercentClass":"success","pendingPercentClass":"danger"},"suites":{"title":"","suites":[{"title":"E2E Tests - Login - Logout Flow (chrome.83_0_4103_106.windows)","suites":[],"tests":[{"title":"User Should not Login with Invalid Credentials","fullTitle":"User Should not Login with Invalid Credentials","timedOut":false,"duration":9018,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"a083d347-0367-4fe7-8173-ac0144faa699","parentUUID":"43f43102-befc-4bdf-9c2c-ae2e685c3aa2","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"3839386ff927ab464dfcd371b2d7ac61\"}]","state":"passed","err":{}},{"title":"User Should Login with Valid Credentials","fullTitle":"User Should Login with Valid Credentials","timedOut":false,"duration":2162,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"8ad453c1-2b21-48de-b469-db39f2be32f3","parentUUID":"43f43102-befc-4bdf-9c2c-ae2e685c3aa2","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"3839386ff927ab464dfcd371b2d7ac61\"}]","state":"passed","err":{}},{"title":"User Should be able to Logout","fullTitle":"User Should be able to Logout","timedOut":false,"duration":306,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"bd92658c-e8d2-4142-9b47-85e58cdadf0f","parentUUID":"43f43102-befc-4bdf-9c2c-ae2e685c3aa2","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"3839386ff927ab464dfcd371b2d7ac61\"}]","state":"passed","err":{}}],"pending":[],"root":false,"fullFile":"","file":"","passes":["a083d347-0367-4fe7-8173-ac0144faa699","8ad453c1-2b21-48de-b469-db39f2be32f3","bd92658c-e8d2-4142-9b47-85e58cdadf0f"],"failures":[],"skipped":[],"duration":11489,"_timeout":0,"uuid":"43f43102-befc-4bdf-9c2c-ae2e685c3aa2","beforeHooks":[],"afterHooks":[]}],"tests":[],"pending":[],"root":true,"fullFile":"","file":"","passes":[],"failures":[],"skipped":[],"duration":0,"_timeout":0,"uuid":"3a979d22-8c39-4deb-bd56-cfd75e65e267","beforeHooks":[],"afterHooks":[]},"copyrightYear":2020}
