+++ mochareports/wdio-0-2-mochawesome-reporter.log
@@ -1,0 +1,1 @@
+{"stats":{"suites":1,"tests":2,"passes":2,"pending":0,"failures":0,"start":"2020-06-22T22:33:01.512Z","end":"2020-06-22T22:33:07.451Z","duration":5939,"testsRegistered":2,"passPercent":100,"pendingPercent":0,"other":0,"hasOther":false,"skipped":0,"hasSkipped":false,"passPercentClass":"success","pendingPercentClass":"danger"},"suites":{"title":"","suites":[{"title":"E2E Tests - Search Functionality (chrome.83_0_4103_106.windows)","suites":[],"tests":[{"title":"Should Load Home Page","fullTitle":"Should Load Home Page","timedOut":false,"duration":3895,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"1adf80b1-d185-471f-aacc-f1adfc6b941f","parentUUID":"6b84450f-7d38-4a35-a194-11a289f7c9b7","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"0793a1686ffefcb3cf76411ef6b2fe1b\"}]","state":"passed","err":{}},{"title":"User Should Submit Search Form","fullTitle":"User Should Submit Search Form","timedOut":false,"duration":1094,"speed":"fast","pass":true,"fail":false,"pending":false,"code":"","isRoot":false,"uuid":"a6fd3a22-3499-41fd-bceb-0482b743e508","parentUUID":"6b84450f-7d38-4a35-a194-11a289f7c9b7","skipped":false,"isHook":false,"context":"[{\"title\":\"Session Id\",\"value\":\"0793a1686ffefcb3cf76411ef6b2fe1b\"}]","state":"passed","err":{}}],"pending":[],"root":false,"fullFile":"","file":"","passes":["1adf80b1-d185-471f-aacc-f1adfc6b941f","a6fd3a22-3499-41fd-bceb-0482b743e508"],"failures":[],"skipped":[],"duration":4993,"_timeout":0,"uuid":"6b84450f-7d38-4a35-a194-11a289f7c9b7","beforeHooks":[],"afterHooks":[]}],"tests":[],"pending":[],"root":true,"fullFile":"","file":"","passes":[],"failures":[],"skipped":[],"duration":0,"_timeout":0,"uuid":"d2c39666-823c-4c31-bc46-0240a79a7f44","beforeHooks":[],"afterHooks":[]},"copyrightYear":2020}
