+++ package.json
@@ -4,7 +4,8 @@
   "description": "Automated Software Testing with WebDriverIO",
   "main": "index.js",
   "scripts": {
-    "test": "wdio wdio.conf.js"
+    "test": "wdio wdio.conf.js",
+    "Login": "wdio wdio.conf.js --suite login"
   },
   "keywords": [
     "Automation",
@@ -17,15 +18,20 @@
     "@babel/core": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
     "@babel/register": "^7.9.0",
+    "@wdio/allure-reporter": "^6.1.14",
     "@wdio/browserstack-service": "^6.1.15",
     "@wdio/cli": "^6.1.12",
+    "@wdio/dot-reporter": "^6.1.14",
     "@wdio/local-runner": "^6.1.12",
     "@wdio/mocha-framework": "^6.1.8",
     "@wdio/spec-reporter": "^6.1.12",
     "@wdio/sync": "^6.1.8",
+    "allure-commandline": "^2.13.0",
     "chromedriver": "^83.0.0",
     "eslint-plugin-mocha": "^7.0.0",
     "prettier": "^2.0.5",
-    "wdio-chromedriver-service": "^6.0.3"
-  }
+    "wdio-chromedriver-service": "^6.0.3",
+    "wdio-mochawesome-reporter": "^4.0.0"
+  },
+  "dependencies": {}
 }
